Restricting export function types and other changes: - Removing extend8 and extend16 instructions - Allowing at most one return type throughout the project - Allowing to specify and restrict the allowed return types for exported functions - Storing indices as usize instead of u32

+++ crates/wasm-encoder/src/functions.rs
@@ -35,8 +35,8 @@ impl FunctionSection {
     }
 
     /// Define a function that uses the given type.
-    pub fn function(&mut self, type_index: u32) -> &mut Self {
-        self.bytes.extend(encoders::u32(type_index));
+    pub fn function(&mut self, type_index: usize) -> &mut Self {
+        self.bytes.extend(encoders::u32(type_index as u32));
         self.num_added += 1;
         self
     }
